refactor docstrings
