+++ feature_engine/creation/cyclical.py
@@ -4,16 +4,16 @@ import numpy as np
 import pandas as pd
 from sklearn.utils import deprecated
 
-from feature_engine._docstrings.class_inputs import (
-    _drop_original_docstring,
-    _variables_numerical_docstring,
-)
 from feature_engine._docstrings.fit_attributes import (
     _feature_names_in_docstring,
     _n_features_in_docstring,
     _variables_attribute_docstring,
 )
 from feature_engine._docstrings.methods import _fit_transform_docstring
+from feature_engine._docstrings.init_parameters import (
+    _variables_numerical_docstring,
+    _drop_original_docstring,
+)
 from feature_engine._docstrings.substitute import Substitution
 from feature_engine._variable_handling.init_parameter_checks import (
     _check_init_parameter_variables,
