+++ feature_engine/creation/mathematical_combination.py
@@ -5,10 +5,6 @@ from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils import deprecated
 from sklearn.utils.validation import check_is_fitted
 
-from feature_engine._docstrings.class_inputs import (
-    _drop_original_docstring,
-    _missing_values_docstring,
-)
 from feature_engine._docstrings.fit_attributes import (
     _feature_names_in_docstring,
     _n_features_in_docstring,
@@ -16,6 +12,10 @@ from feature_engine._docstrings.fit_attributes import (
 from feature_engine._docstrings.methods import (
     _fit_not_learn_docstring,
     _fit_transform_docstring,
+)
+from feature_engine._docstrings.init_parameters import (
+    _drop_original_docstring,
+    _missing_values_docstring,
 )
 from feature_engine._docstrings.substitute import Substitution
 from feature_engine._variable_handling.variable_type_selection import (
