+++ feature_engine/encoding/probability_ratio.py
@@ -14,14 +14,14 @@ from feature_engine._docstrings.fit_attributes import (
 from feature_engine._docstrings.methods import (
     _fit_transform_docstring,
     _inverse_transform_docstring,
+    _transform_encoders_docstring,
 )
 from feature_engine._docstrings.substitute import Substitution
 from feature_engine.dataframe_checks import check_X_y
-from feature_engine.encoding._docstrings import (
+from feature_engine._docstrings.init_parameters import (
+    _variables_categorical_docstring,
     _ignore_format_docstring,
-    _transform_docstring,
     _unseen_docstring,
-    _variables_docstring,
 )
 from feature_engine.encoding.base_encoder import (
     CategoricalInitExpandedMixin,
@@ -32,13 +32,13 @@ from feature_engine.tags import _return_tags
 
 @Substitution(
     ignore_format=_ignore_format_docstring,
-    variables=_variables_docstring,
+    variables=_variables_categorical_docstring,
     unseen=_unseen_docstring,
     variables_=_variables_attribute_docstring,
     feature_names_in_=_feature_names_in_docstring,
     n_features_in_=_n_features_in_docstring,
     fit_transform=_fit_transform_docstring,
-    transform=_transform_docstring,
+    transform=_transform_encoders_docstring,
     inverse_transform=_inverse_transform_docstring,
 )
 class PRatioEncoder(CategoricalInitExpandedMixin, CategoricalMethodsMixin):
