+++ feature_engine/timeseries/forecasting/expanding_window_features.py
@@ -6,11 +6,7 @@ from __future__ import annotations
 import pandas as pd
 from sklearn.utils.validation import check_is_fitted
 
-from feature_engine._docstrings.class_inputs import (
-    _drop_original_docstring,
-    _missing_values_docstring,
-    _variables_numerical_docstring,
-)
+
 from feature_engine._docstrings.fit_attributes import (
     _feature_names_in_docstring,
     _n_features_in_docstring,
@@ -18,6 +14,11 @@ from feature_engine._docstrings.fit_attributes import (
 from feature_engine._docstrings.methods import (
     _fit_not_learn_docstring,
     _fit_transform_docstring,
+)
+from feature_engine._docstrings.init_parameters import (
+    _variables_numerical_docstring,
+    _drop_original_docstring,
+    _missing_values_docstring,
 )
 from feature_engine._docstrings.substitute import Substitution
 from feature_engine.timeseries.forecasting.base_forecast_transformers import (
