+++ feature_engine/timeseries/forecasting/window_features.py
@@ -3,11 +3,6 @@ from typing import Callable, List, Optional, Union
 import pandas as pd
 from sklearn.utils.validation import check_is_fitted
 
-from feature_engine._docstrings.class_inputs import (
-    _drop_original_docstring,
-    _missing_values_docstring,
-    _variables_numerical_docstring,
-)
 from feature_engine._docstrings.fit_attributes import (
     _feature_names_in_docstring,
     _n_features_in_docstring,
@@ -15,6 +10,11 @@ from feature_engine._docstrings.fit_attributes import (
 from feature_engine._docstrings.methods import (
     _fit_not_learn_docstring,
     _fit_transform_docstring,
+)
+from feature_engine._docstrings.init_parameters import (
+    _variables_numerical_docstring,
+    _drop_original_docstring,
+    _missing_values_docstring,
 )
 from feature_engine._docstrings.substitute import Substitution
 from feature_engine.timeseries.forecasting.base_forecast_transformers import (
