+++ feature_engine/transformation/boxcox.py
@@ -7,7 +7,7 @@ import pandas as pd
 import scipy.stats as stats
 from scipy.special import inv_boxcox
 
-from feature_engine._docstrings.class_inputs import _variables_numerical_docstring
+from feature_engine._docstrings.init_parameters import _variables_numerical_docstring
 from feature_engine._docstrings.fit_attributes import (
     _feature_names_in_docstring,
     _n_features_in_docstring,
