+++ feature_engine/transformation/power.py
@@ -6,17 +6,18 @@ from typing import List, Optional, Union
 import numpy as np
 import pandas as pd
 
-from feature_engine._docstrings.class_inputs import _variables_numerical_docstring
-from feature_engine._docstrings.fit_attributes import (
-    _feature_names_in_docstring,
-    _n_features_in_docstring,
-    _variables_attribute_docstring,
-)
 from feature_engine._docstrings.methods import (
     _fit_not_learn_docstring,
     _fit_transform_docstring,
     _inverse_transform_docstring,
 )
+
+from feature_engine._docstrings.fit_attributes import (
+    _variables_attribute_docstring,
+    _feature_names_in_docstring,
+    _n_features_in_docstring,
+)
+from feature_engine._docstrings.init_parameters import _variables_numerical_docstring
 from feature_engine._docstrings.substitute import Substitution
 from feature_engine._variable_handling.init_parameter_checks import (
     _check_init_parameter_variables,
